+++ js/protobuf/state-update.js
@@ -3,11 +3,11 @@ const ViewUpdateStuff = {
   Turtle: {
     fields: {
       "breed":       { type: "string", id:  1 }
-    , "color":       { type: "double", id:  2 }
+    , "color":       { type: "uint32", id:  2 }
     , "heading":     { type: "double", id:  3 }
     , "hidden?":     { type: "bool"  , id:  4 }
     , "label":       { type: "string", id:  5 }
-    , "label-color": { type: "double", id:  6 }
+    , "label-color": { type: "uint32", id:  6 }
     , "pen-mode":    { type: "string", id:  7 }
     , "pen-size":    { type: "double", id:  8 }
     , "shape":       { type: "string", id:  9 }
@@ -20,9 +20,9 @@ const ViewUpdateStuff = {
 
 , Patch: {
     fields: {
-      "pcolor":       { type: "double", id: 1 }
+      "pcolor":       { type: "uint32", id: 1 }
     , "plabel":       { type: "string", id: 2 }
-    , "plabel-color": { type: "double", id: 3 }
+    , "plabel-color": { type: "uint32", id: 3 }
     , "pxcor":        { type: "double", id: 4 }
     , "pycor":        { type: "double", id: 5 }
     }
@@ -31,14 +31,14 @@ const ViewUpdateStuff = {
 , Link: {
     fields: {
       "breed":       { type: "string", id:  1 }
-    , "color":       { type: "double", id:  2 }
+    , "color":       { type: "uint32", id:  2 }
     , "end1":        { type: "uint32", id:  3 }
     , "end2":        { type: "uint32", id:  4 }
     , "heading":     { type: "double", id:  5 }
     , "hidden?":     { type: "bool"  , id:  6 }
     , "directed?":   { type: "bool"  , id:  7 }
     , "label":       { type: "string", id:  8 }
-    , "label-color": { type: "double", id:  9 }
+    , "label-color": { type: "uint32", id:  9 }
     , "midpointx":   { type: "double", id: 10 }
     , "midpointy":   { type: "double", id: 11 }
     , "shape":       { type: "string", id: 12 }
@@ -166,7 +166,7 @@ const PlotUpdateStuff = {
   Pen: {
     fields: {
       name:            { type: "string", id: 1 }
-    , color:           { type: "double", id: 2 }
+    , color:           { type: "uint32", id: 2 }
     }
   }
 
@@ -208,7 +208,7 @@ const PlotUpdateStuff = {
 , PlotUpdatePenColor: {
     fields: {
       penName: { type: "string", id: 1 }
-    , color:   { type: "double", id: 2 }
+    , color:   { type: "uint32", id: 2 }
     }
   }
